Renamed Textbit to TextbitRoot. exported as Root
