--- src/components/Textbit/index.tsx
@@ -1,2 +1,0 @@
-export { Textbit } from './Textbit'
-export { useTextbit } from './useTextbit'
