+++ src/components/TextbitEditable/TextbitEditable.tsx
@@ -25,7 +25,7 @@ import { withEditableVoids } from './with/editableVoids'
 import { withInsertBreak } from './with/insertBreak'
 import { withInsertHtml } from './with/insertHtml'
 import { PresenceOverlay } from './components/PresenceOverlay/PresenceOverlay'
-import { useTextbit } from '../Textbit'
+import { useTextbit } from '../TextbitRoot'
 import { debounce } from '@/lib/debounce'
 import { usePluginRegistry } from '../PluginRegistry'
 import { PluginRegistryAction, PluginRegistryComponent } from '../PluginRegistry/lib/types'
