+++ src/components/TextbitRoot/index.tsx
@@ -1,0 +1,2 @@
+export { TextbitRoot } from './TextbitRoot'
+export { useTextbit } from './useTextbit'
